Refactoring entire backend.

+++ src/main/java/com/example/idatt2106_2022_05_backend/controller/UserController.java
@@ -2,13 +2,14 @@ package com.example.idatt2106_2022_05_backend.controller;
 
 import com.example.idatt2106_2022_05_backend.dto.user.UserReturnDto;
 import com.example.idatt2106_2022_05_backend.dto.user.UserUpdateDto;
+import com.example.idatt2106_2022_05_backend.security.SecurityService;
 import com.example.idatt2106_2022_05_backend.service.user.UserService;
 import com.example.idatt2106_2022_05_backend.util.Response;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import lombok.extern.slf4j.Slf4j;
-import org.apache.tomcat.util.http.parser.Authorization;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.web.bind.annotation.*;
@@ -24,10 +25,16 @@ public class UserController {
     @Autowired
     private UserService userService;
 
+    @Autowired
+    private SecurityService securityService;
+
     @PutMapping("/{userId}")
     @ApiOperation(value = "Endpoint to update user", response = Response.class)
     public Response update(@PathVariable Long userId, @RequestBody UserUpdateDto userUpdateDto) throws IOException {
         log.debug("[X] Call to update user with id = {}", userId);
+        if(!securityService.isUser(userId)){
+            return new Response("Du har ikke tilgang", HttpStatus.BAD_REQUEST);
+        }
         return userService.updateUser(userId, userUpdateDto);
     }
 
@@ -35,6 +42,9 @@ public class UserController {
     @ApiOperation(value = "Endpoint to delete user", response = Response.class)
     public Response deleteUser(@PathVariable Long userId) {
         log.debug("[X] Call to delete user with id = {}", userId);
+        if(!securityService.isUser(userId)){
+            return new Response("Du har ikke tilgang", HttpStatus.BAD_REQUEST);
+        }
         return userService.deleteUser(userId);
     }
